requirements.txt
requests>=2.31.0
loguru>=0.7.2
playwright>=1.40.0
pyyaml>=6.0